+++ resolwe/flow/models/annotations.py
@@ -17,16 +17,22 @@ from typing import (
     Type,
 )
 
+from django.conf import settings
 from django.core.exceptions import ValidationError
 from django.db import models
+from versionfield import VersionField
 
 from resolwe.flow.models.base import BaseModel
-from resolwe.permissions.models import PermissionInterface, PermissionObject
+from resolwe.permissions.models import (
+    PermissionInterface,
+    PermissionObject,
+    PermissionQuerySet,
+)
 
 if TYPE_CHECKING:
     from resolwe.flow.models import Collection, Entity
 
-from .base import AuditModel
+from .base import VERSION_NUMBER_BITS, AuditModel
 
 VALIDATOR_LENGTH = 128
 NAME_LENGTH = 128
@@ -234,6 +240,28 @@ class AnnotationGroup(models.Model):
         ordering = ["sort_order"]
 
 
+class AnnotationFieldManager(models.Manager):
+    """Annotation field manager.
+
+    Return only the latest version of each field.
+    """
+
+    def get_queryset(self) -> models.QuerySet:
+        """Return only the latest version for every field."""
+        return (
+            super()
+            .get_queryset()
+            .annotate(
+                rank=models.Window(
+                    expression=models.functions.DenseRank(),
+                    partition_by=[models.F("name"), models.F("group")],
+                    order_by=models.F("version").desc(),
+                ),
+            )
+            .filter(rank=1)
+        )
+
+
 class AnnotationField(models.Model):
     """Annotation field."""
 
@@ -267,6 +295,11 @@ class AnnotationField(models.Model):
 
     #: is this field required
     required = models.BooleanField(default=False)
+
+    #: field version
+    version = VersionField(number_bits=VERSION_NUMBER_BITS, default="0.0.0")
+
+    objects = AnnotationFieldManager()
 
     def __init__(self, *args, **kwargs):
         """Store original vocabulary to private variable."""
@@ -329,7 +362,7 @@ class AnnotationField(models.Model):
                 annotation_value_field.recompute_label()
                 annotation_value_field.validate()
                 updated_fields.append(annotation_value_field)
-            AnnotationValue.objects.bulk_update(updated_fields, ["_value"])
+            AnnotationValue.all_objects.bulk_update(updated_fields, ["_value"])
             self._original_vocabulary = self.vocabulary
         if self.type != self._original_type:
             self.revalidate_values()
@@ -368,7 +401,8 @@ class AnnotationField(models.Model):
                 name="annotation_type",
             ),
             models.constraints.UniqueConstraint(
-                fields=["name", "group"], name="uniquetogether_name_group"
+                fields=["name", "group", "version"],
+                name="uniquetogether_name_group_version",
             ),
         ]
         ordering = ["group__sort_order", "sort_order"]
@@ -395,6 +429,35 @@ class AnnotationPreset(BaseModel, PermissionObject):
         """Override parent meta."""
 
 
+class AnnotationValueManager(
+    models.manager.BaseManager.from_queryset(PermissionQuerySet)
+):
+    """Get the latest value by default."""
+
+    def get_queryset(self) -> models.QuerySet:
+        """Return the latest version for every value."""
+        print("Get av qset")
+        return self._only_latest(super().get_queryset()).exclude(
+            _value__value=models.Value(None, models.JSONField())
+        )
+
+    def _only_latest(
+        self, queryset: models.QuerySet["AnnotationValue"]
+    ) -> models.QuerySet["AnnotationValue"]:
+        """Return only latest values from the queryset."""
+        return queryset.annotate(
+            rank=models.Window(
+                expression=models.functions.DenseRank(),
+                partition_by=[models.F("field"), models.F("entity")],
+                order_by=models.F("created").desc(),
+            ),
+        ).filter(rank=1)
+
+    def history(self, timestamp: datetime.datetime):
+        """Get values at certain point in time"""
+        return self._only_latest(self.filter(created__le=timestamp))
+
+
 class AnnotationValue(PermissionInterface, AuditModel):
     """The value of the annotation."""
 
@@ -403,7 +466,8 @@ class AnnotationValue(PermissionInterface, AuditModel):
 
         constraints = [
             models.constraints.UniqueConstraint(
-                fields=["entity", "field"], name="uniquetogether_entity_field"
+                fields=["entity", "field", "created"],
+                name="uniquetogether_entity_field_created",
             ),
         ]
         ordering = ["field__group__sort_order", "field__sort_order"]
@@ -418,11 +482,18 @@ class AnnotationValue(PermissionInterface, AuditModel):
         AnnotationField, related_name="values", on_delete=models.PROTECT
     )
 
-    #: the date when field was last modified
-    modified = models.DateTimeField(auto_now=True)
+    #: the date when field was created
+    created = models.DateTimeField(auto_now_add=True)
 
     #: value is stored under key 'value' in the json field to simplify lookups
     _value: Any = models.JSONField(default=dict)
+
+    #: user that created the entry
+    contributor = models.ForeignKey(settings.AUTH_USER_MODEL, on_delete=models.PROTECT)
+
+    objects = AnnotationValueManager()
+
+    all_objects = PermissionQuerySet.as_manager()
 
     def __init__(self, *args, **kwargs):
         """Allow us to set the 'value' in the constructor.
@@ -477,7 +548,9 @@ class AnnotationValue(PermissionInterface, AuditModel):
 
         :raises ValidationError: when the validation fails.
         """
-        annotation_value_validator.validate(self, raise_exception=True)
+        # The value `None` is used as delete marker.
+        if self.value is not None:
+            annotation_value_validator.validate(self, raise_exception=True)
 
     def save(self, *args, **kwargs):
         """Save the annotation value after validation has passed."""