+++ resolwe/flow/models/entity.py
@@ -349,7 +349,7 @@ class Entity(BaseCollection, PermissionObject):
         if validation_errors:
             raise ValidationError(validation_errors)
 
-    def update_annotations(self, annotations: dict[str, Any], update=True):
+    def update_annotations(self, annotations: dict[str, Any], contributor, update=True):
         """Update annotations with the given values.
 
         When annotation value is set no None it is deleted.
@@ -367,34 +367,26 @@ class Entity(BaseCollection, PermissionObject):
         # be deleted.
         annotation_values: list[AnnotationValue] = []
         validation_errors: list[ValidationError] = []
-        fields_to_delete: list[int] = []
         for field_path, field_value in annotations.items():
-            if field_value is None:
-                fields_to_delete.append(field_map[field_path].pk)
-            else:
-                value = AnnotationValue(
-                    field=field_map[field_path], value=field_value, entity=self
-                )
-                annotation_values.append(value)
-                try:
-                    value.validate()
-                except ValidationError as e:
-                    validation_errors.append(e)
+            value = AnnotationValue(
+                field=field_map[field_path],
+                value=field_value,
+                entity=self,
+                contributor=contributor,
+            )
+            annotation_values.append(value)
+            try:
+                value.validate()
+            except ValidationError as e:
+                validation_errors.append(e)
         if validation_errors:
             raise ValidationError(validation_errors)
 
-        # Delete and update annotations in a transaction.
+
+        print("Creating", annotation_values)
+        # Create new entries in a transaction.
         with transaction.atomic():
-            to_delete = AnnotationValue.objects.filter(entity=self)
-            if update:
-                to_delete = to_delete.filter(field_id__in=fields_to_delete)
-            to_delete.delete()
-            AnnotationValue.objects.bulk_create(
-                annotation_values,
-                update_conflicts=True,
-                update_fields=["_value"],
-                unique_fields=["entity", "field"],
-            )
+            AnnotationValue.objects.bulk_create(annotation_values)
 
         # Add missing annotation fields to the collection.
         if self.collection is not None: